pyproject.toml
[tool.poetry]
name = "tft-backend"
version = "0.1.0"
description = ""
authors = ["siggu <siggu100@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
pillow = "^10.2.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
